moved some functions to util so they are accessible from util
